Refactored benchmark folder structure
