+++ topic_benchmark/benchmark/__init__.py
@@ -1,0 +1,3 @@
+from topic_benchmark.benchmark.base import (BenchmarkEntry, BenchmarkError,
+                                            EnryID)
+from topic_benchmark.benchmark.benchmark import run_benchmark
